+++ lib/src/text_analyzer/_index.dart
@@ -2,9 +2,6 @@
 // Copyright ©2022, GM Consult Pty Ltd
 // All rights reserved
 
-export 'n_gram_range.dart';
+export 'co_occurrence_graph.dart';
 export 'text_analysis_extensions.dart';
 export 'text_analyzer.dart';
-export 'text_document.dart';
-export 'text_tokenizer.dart';
-export 'token.dart';