+++ lib/src/text_document/_index.dart
@@ -1,0 +1,5 @@
+// BSD 3-Clause License
+// Copyright ©2022, GM Consult Pty Ltd
+// All rights reserved
+
+export 'text_document.dart';